Process: com.mario_antolovic.weatherappupgraded, PID: 8934     java.lang.IllegalArgumentException: Target must not be null.     got error with the image view view trying to solve it     Other things are completed

+++ app/src/main/java/com/mario_antolovic/weatherappupgraded/Model/Main.java
@@ -3,7 +3,7 @@ package com.mario_antolovic.weatherappupgraded.Model;
 public class Main {
 
     private double temp ;
-    private int pressure ;
+    private float pressure ;
     private int humidity ;
     private double temp_min ;
     private double temp_max ;
@@ -21,11 +21,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
